minor: refactoring, docstring
